Added generic dR-based overlap removal method.
The method will take an object and a container and return true if the
object overlaps with any object in the container (except itself) within
some specified dR cone. It's not obvious that such a method should be
used, since it can't be used in all cases and complicates the code
slightly. Should readability be the ultimate goal, or is minimizing code
duplication more important?