Added alternate method for tau-lep OR.
For the case of using view-containers or subset containers with the
tool, it is necessary to be able to pass in the loose electron and muon
containers for the tau-e/mu OR. To this end, I provided an alternate
top-level method which consumes an additional electron and muon
container. Now, the original method calls the new one with the electrons
and muons passed as both sets of containers.